+++ plugin-route-collectors/src/testFixtures/kotlin/com/linecorp/intellij/plugins/armeria/test/ArmeriaFixtureScenarioStubs.kt
@@ -150,6 +150,7 @@ fun JavaCodeInsightTestFixture.registerRouteDuplicateIndexStubs() {
         }
         """.trimIndent(),
     )
+    registerKnownHttpServiceStubs()
 }
 
 fun JavaCodeInsightTestFixture.registerRouteCollectorStubs() {
@@ -242,50 +243,7 @@ fun JavaCodeInsightTestFixture.registerRouteCollectorStubs() {
         }
         """.trimIndent(),
     )
-    this.addClass(
-        """
-        package com.linecorp.armeria.server.grpc;
-
-        public final class GrpcService {
-            public static GrpcServiceBuilder builder(Object bindableService) {
-                return null;
-            }
-        }
-        """.trimIndent(),
-    )
-    this.addClass(
-        """
-        package com.linecorp.armeria.server.grpc;
-
-        public final class GrpcServiceBuilder {
-            public com.linecorp.armeria.server.grpc.GrpcService build() {
-                return null;
-            }
-        }
-        """.trimIndent(),
-    )
-    this.addClass(
-        """
-        package com.linecorp.armeria.server.docs;
-
-        public final class DocService {
-            public static DocServiceBuilder builder() {
-                return null;
-            }
-        }
-        """.trimIndent(),
-    )
-    this.addClass(
-        """
-        package com.linecorp.armeria.server.docs;
-
-        public final class DocServiceBuilder {
-            public com.linecorp.armeria.server.docs.DocService build() {
-                return null;
-            }
-        }
-        """.trimIndent(),
-    )
+    registerKnownHttpServiceStubs()
 }
 
 fun JavaCodeInsightTestFixture.registerKotlinRouteCollectorStubs() {
@@ -392,48 +350,5 @@ fun JavaCodeInsightTestFixture.registerKotlinRouteCollectorStubs() {
         }
         """.trimIndent(),
     )
-    this.addClass(
-        """
-        package com.linecorp.armeria.server.grpc;
-
-        public final class GrpcService {
-            public static GrpcServiceBuilder builder(Object bindableService) {
-                return null;
-            }
-        }
-        """.trimIndent(),
-    )
-    this.addClass(
-        """
-        package com.linecorp.armeria.server.grpc;
-
-        public final class GrpcServiceBuilder {
-            public com.linecorp.armeria.server.grpc.GrpcService build() {
-                return null;
-            }
-        }
-        """.trimIndent(),
-    )
-    this.addClass(
-        """
-        package com.linecorp.armeria.server.docs;
-
-        public final class DocService {
-            public static DocServiceBuilder builder() {
-                return null;
-            }
-        }
-        """.trimIndent(),
-    )
-    this.addClass(
-        """
-        package com.linecorp.armeria.server.docs;
-
-        public final class DocServiceBuilder {
-            public com.linecorp.armeria.server.docs.DocService build() {
-                return null;
-            }
-        }
-        """.trimIndent(),
-    )
+    registerKnownHttpServiceStubs()
 }